// frontend/cms-frontend-ts/src/api/generated/api.ts
export interface StudentResponse {
    'id'?: number;
    'studentcode'?: string;
    'fullname'?: string;
    'dateofbirth'?: string;
    'email'?: string;
    'phone'?: string;
    'address'?: string;
    'major'?: string;
    'classname'?: string;
    'createdAt'?: string;
    'updatedAt'?: string;
}